refactor: split code into mobile, web, and common view/components
