--- src/components/ChangePasswordModal.vue
@@ -1,19 +1,0 @@
-<template>
-  <ion-toolbar>
-    <ion-title> Change Password </ion-title>
-    <ion-buttons slot="primary">
-      <ion-button>Save</ion-button>
-    </ion-buttons>
-    <ion-buttons slot="secondary">
-      <ion-button>Cancel</ion-button>
-    </ion-buttons>
-  </ion-toolbar>
-
-  <ion-content> placeholder </ion-content>
-</template>
-
-<script setup lang="ts">
-import { IonToolbar, IonTitle, IonButtons, IonButton, IonContent } from '@ionic/vue';
-</script>
-
-<style scoped></style>
