--- src/components/SettingsModal.vue
@@ -1,10 +1,0 @@
-<template>
-  <ion-nav id="settings-nav" :root="SettingsNavRoot"> </ion-nav>
-</template>
-
-<script setup lang="ts">
-import { IonNav } from '@ionic/vue';
-import SettingsNavRoot from './SettingsNavRoot.vue';
-</script>
-
-<style scoped></style>
