--- src/components/YouAccountModal.vue
@@ -1,10 +1,0 @@
-<template>
-  <ion-nav id="account-nav" :root="YouAccountModalRoot"> </ion-nav>
-</template>
-
-<script setup lang="ts">
-import { IonNav } from '@ionic/vue';
-import YouAccountModalRoot from './YouAccountModalRoot.vue';
-</script>
-
-<style scoped></style>
