+++ src/components/common/PlaceholderComponent.vue
