+++ src/components/mobile/AccountModal.vue
@@ -1,0 +1,10 @@
+<template>
+  <ion-nav id="account-nav" :root="AccountRoot"> </ion-nav>
+</template>
+
+<script setup lang="ts">
+import { IonNav } from '@ionic/vue';
+import AccountRoot from '@components/mobile/AccountRoot.vue';
+</script>
+
+<style scoped></style>
