+++ src/components/mobile/SettingsModal.vue
@@ -1,0 +1,10 @@
+<template>
+  <ion-nav id="settings-nav" :root="SettingsRoot"> </ion-nav>
+</template>
+
+<script setup lang="ts">
+import { IonNav } from '@ionic/vue';
+import SettingsRoot from '@components/mobile/SettingsRoot.vue';
+</script>
+
+<style scoped></style>
