+++ src/composables/modalControls.ts
@@ -1,26 +1,27 @@
+/* eslint-disable @typescript-eslint/no-explicit-any */
 import { modalController } from '@ionic/vue';
 
 export { useModalControls };
 
-function useModalControls(
-  startingBreakPoint: number,
-  breakpointList: Array<number>,
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  content: any,
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  contentProps?: { [key: string]: any },
-  presenter?: HTMLElement
-) {
-  const initialBreakPoint = startingBreakPoint;
-  const breakPoints = breakpointList;
+type modalConfig = {
+  startingBreakPoint?: number;
+  breakpointList?: number[];
+  content: any;
+  contentProps?: { [key: string]: any };
+  presentingElement?: HTMLElement;
+};
+
+function useModalControls(config: modalConfig) {
+  const initialBreakPoint = config.startingBreakPoint;
+  const breakPoints = config.breakpointList;
 
   async function show() {
     const modal = await modalController.create({
-      breakpoints: breakpointList,
-      initialBreakpoint: startingBreakPoint,
-      component: content,
-      componentProps: contentProps,
-      presentingElement: presenter
+      breakpoints: config.breakpointList,
+      initialBreakpoint: config.startingBreakPoint,
+      component: config.content,
+      componentProps: config.contentProps,
+      presentingElement: config.presentingElement
     });
 
     modal.present();
