+++ src/views/common/HomeView.vue
@@ -1,3 +1,20 @@
+<template>
+  <ion-page>
+    <ion-header class="ion-no-border">
+      <ion-toolbar>
+        <ion-title class="ion-text-center" size="large">REMEMBIT</ion-title>
+      </ion-toolbar>
+    </ion-header>
+    <ion-content class="ion-padding">
+      <p>
+        This is a personal project for learning. The goal is a simple and cross-platform
+        bookmarking tool for those of us who use multiple browsers.
+      </p>
+      <ion-button @click="router.push({ name: 'Login' })">Log In</ion-button>
+    </ion-content>
+  </ion-page>
+</template>
+
 <script setup lang="ts">
 import {
   IonPage,
@@ -11,22 +28,5 @@ import { useRouter } from 'vue-router';
 
 const router = useRouter();
 </script>
-
-<template>
-  <ion-page>
-    <ion-header class="ion-no-border">
-      <ion-toolbar>
-        <ion-title class="ion-text-center" size="large">REMEMBIT</ion-title>
-      </ion-toolbar>
-    </ion-header>
-    <ion-content class="ion-padding">
-      <p>
-        This is a personal project for learning. The goal is a simple and cross-platform
-        bookmarking tool for those of us who use multiple browsers.
-      </p>
-      <ion-button @click="router.push({ name: 'Bookmarks' })">Log In</ion-button>
-    </ion-content>
-  </ion-page>
-</template>
 
 <style scoped></style>
